+++ docs/source/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Latest
+## 0.6.2 (2023-07-23)
 - Switch back to using numpy nullable dtypes for Pandas as default.
 - Ensure that indices get created in same schema as corresponding table (IBM Db2).
 - Fix private method `SQLTableStore.get_stage_hash`.